:bug::recycle::iphone: bug with website fixed. EnterPage refactured and media for mobile updated
