+++ public/index.html
@@ -6,45 +6,30 @@
     <meta name="theme-color" content="#000000" />
     <meta
       name="description"
-      content="Web site created using create-react-app"
-    />
-
-    <title>Concert's Database</title>
-
-    <!-- Information for Card when Website is shared -->
-    <meta
-      property="og:title"
-      content="Project #2: React Application - Concert's Database"
-    />
-
-    <meta
       property="og:description"
-      content="Concert's Database React application developed at Ironhack Web Development Bootcamp, using React, JavaScript, CSS, Material UI and Canvas."
+      content="Website created to simulate a personal concert database, inspired by my passion for music and live concerts. The goal is to keep track of the shows I've been to."
     />
+    <meta name="title" property="og:title" content="Concert's Database" />
     <meta property="og:type" content="article" />
     <meta property="og:url" content="https://concert-database.netlify.app/" />
-    <meta property="og:image" content=".src/images/home-page.png" />
     <meta name="author" content="Tulio Minini and Lucas Colombo" />
+    <meta property="og:image" content="./images/meta-tag-img.jpg" />
 
+    <!-- Google Fonts -->
     <link rel="preconnect" href="https://fonts.googleapis.com" />
     <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
     <link
       href="https://fonts.googleapis.com/css2?family=The+Nautigal&display=swap"
       rel="stylesheet"
     />
+
+    <!-- Favicon -->
+    <link rel="icon" type="image/x-icon" href="favicon.ico" />
+
+    <!-- Web Page Tab -->
+    <title>Concert's Database</title>
   </head>
   <body>
-    <noscript>You need to enable JavaScript to run this app.</noscript>
     <div id="root"></div>
-    <!--
-      This HTML file is a template.
-      If you open it directly in the browser, you will see an empty page.
-
-      You can add webfonts, meta tags, or analytics to this file.
-      The build step will place the bundled scripts into the <body> tag.
-
-      To begin the development, run `npm start` or `yarn start`.
-      To create a production bundle, use `npm run build` or `yarn build`.
-    -->
   </body>
 </html>
