+++ src/index.css
@@ -1,0 +1,3 @@
+body {
+    background-color: #010102;
+  }
