+++ src/pages/EnterPage/EnterPage.module.css
@@ -1,163 +1,146 @@
-body {
-  background-color: #010102;
-}
 .general {
-  height: 740px;
-  width: 1110px;
+  background-image: url("./../../images/crowd-background-2.avif");
+  height: 93vh;
+  max-width: 80vw;
+  padding: 20px 0;
   text-align: center;
   margin: auto;
-  object-fit: cover;
   background-repeat: no-repeat;
-  background-size: contain;
+  background-size: cover;
   display: flex;
   flex-direction: column;
-}
-.enterPage {
-  width: 100%;
-  display: flex;
-  justify-content: center;
-  flex-direction: column;
-  align-items: center;
-}
-.h1 {
-  font-family: "The Nautigal", cursive;
-  font-size: 300px;
-  height: 150px;
-  color: aliceblue;
-  text-align: center;
-  margin-top: 50px;
-  margin-bottom: 0px;
-  margin-left: 0px;
-}
-.h2 {
+  justify-content: space-between;
   font-family: Verdana, Geneva, Tahoma, sans-serif;
-  color: aliceblue;
-  font-size: 25px;
-  text-align: center;
-  height: 30px;
-  padding: 5px 8px;
-  margin-top: 110px;
-  margin-bottom: 0px;
-  margin-left: 450px;
 }
-.enterLink {
+
+.enterPage h1 {
+  font-family: "The Nautigal", cursive;
+  font-size: 20rem;
+  letter-spacing: 10px;
+  height: 18.7rem;
+  color: aliceblue;
+  margin: 0;
+}
+
+.enterPage h2 {
+  color: aliceblue;
+  font-size: 2rem;
+  text-align: right;
+  height: 30px;
+  margin-right: 25%;
+  margin-top: -10px;
+}
+
+.enterPage p {
   color: aliceblue;
   margin: 0px;
-  margin-top: 60px;
-  font-size: 50px;
-}
-.createdBy {
-  font-family: Verdana, Geneva, Tahoma, sans-serif;
-  display: flex;
-  flex-direction: column;
-  justify-content: center;
-  align-items: center;
   margin-top: 150px;
+  font-size: 3.125rem;
 }
+
 .createdBy h2 {
   color: aliceblue;
-  font-size: 14px;
-  letter-spacing: 2px;
+  font-size: 0.95rem;
+  letter-spacing: 1px;
   margin-bottom: 0px;
 }
 .namesGroup {
-  text-align: center;
   color: aliceblue;
   display: flex;
   justify-content: center;
-  align-items: center;
 }
 .names {
   margin: 0px 10px;
 }
+
 .names h3 {
-  font-size: 20px;
+  font-size: 1.25rem;
+  letter-spacing: 2px;
   margin: 15px 0px 0px 0px;
 }
 
-@media only screen and (max-width: 950px) {
-  body {
-    background-color: #010102;
-  }
+/* iPad Air 820 x 1180  */
+
+@media (max-width: 820px) {
   .general {
-    height: 1165px;
-    width: 100%;
-    object-fit: cover;
-    background-repeat: no-repeat;
-    background-size: 820px;
-    display: flex;
-    flex-direction: column;
+    max-height: 100vh;
+    max-width: 100vw;
+    padding: 60px 0 20px;
   }
-  .h1 {
-    font-family: "The Nautigal", cursive;
-    font-size: 280px;
-    color: aliceblue;
-    margin-top: 150px;
-    margin-right: 40px;
+
+  .enterPage h1 {
+    font-size: 16rem;
+    letter-spacing: 10px;
+    height: 15rem;
   }
-  .h2 {
-    font-family: Verdana, Geneva, Tahoma, sans-serif;
-    color: aliceblue;
-    font-size: 30px;
-    margin-top: 93px;
-    margin-left: 340px;
+
+  .enterPage h2 {
+    font-size: 2rem;
+    margin-right: 15%;
   }
-  .enterLink {
-    color: aliceblue;
-    margin: 0px;
-    margin-top: 210px;
-    margin-bottom: 170px;
-    font-size: 60px;
-  }
-  .createdBy h2 {
-    letter-spacing: 2px;
-    font-size: 14px;
+  .enterPage p {
+    margin-top: 270px;
   }
 }
 
-@media only screen and (max-width: 440px) {
-  body {
-    background-color: #010102;
-  }
+/* iPhone 11 Pro Max / XR 414 x 896  */
+
+@media (max-width: 414px) {
   .general {
-    height: 827px;
-    width: 100%;
-    object-fit: cover;
-    background-repeat: no-repeat;
-    background-size: 700px;
-    display: flex;
-    flex-direction: column;
+    max-height: 89vh;
   }
 
-  .h1 {
-    font-family: "The Nautigal", cursive;
-    font-size: 150px;
-    height: 120px;
-    color: aliceblue;
-    margin-top: 100px;
-    margin-right: 25px;
-  }
-  .h2 {
-    font-family: Verdana, Geneva, Tahoma, sans-serif;
-    color: aliceblue;
-    font-size: 20px;
-    margin-top: 5px;
-    margin-left: 160px;
-  }
-  .enterLink {
-    color: aliceblue;
-    margin: 0px;
-    margin-top: 150px;
-    margin-bottom: 110px;
-    font-size: 40px;
+  .enterPage h1 {
+    font-size: 8.5rem;
+    letter-spacing: 7px;
+    margin-top: 70px;
+    height: 8.5rem;
   }
 
-  .createdBy h2 {
-    font-size: 10px;
-    letter-spacing: 2px;
+  .enterPage h2 {
+    font-size: 1.2rem;
+    margin-right: 15%;
   }
+
+  .enterPage p {
+    margin-top: 200px;
+    font-size: 2.5rem;
+  }
+
   .names h3 {
-    font-size: 15px;
+    font-size: 0.9rem;
+    letter-spacing: 2px;
     margin: 15px 0px 0px 0px;
+  }
+}
+
+/* iPhone 12 Pro / 13 Pro 390 x 844  */
+
+@media (max-width: 390px) {
+  .general {
+    max-height: 88vh;
+  }
+
+  .enterPage h1 {
+    /* font-size: 8.5rem; */
+    /* letter-spacing: 7px; */
+    margin-top: 60px;
+    /* height: 8.5rem; */
+  }
+
+  .enterPage h2 {
+    /* font-size: 1.2rem; */
+    /* margin-right: 15%; */
+  }
+
+  .enterPage p {
+    margin-top: 180px;
+    font-size: 2.1rem;
+  }
+
+  .names h3 {
+    /* font-size: .9rem;
+    letter-spacing: 2px;
+    margin: 15px 0px 0px 0px; */
   }
 }
